+++ tests/Integration/AbstractCapiBouncerTest.php
@@ -5,14 +5,12 @@ declare(strict_types=1);
 namespace CrowdSecBouncer\Tests\Integration;
 
 use CrowdSec\CapiClient\Storage\FileStorage;
-use CrowdSec\CapiClient\Watcher as CapiClient;
 use CrowdSec\Common\Logger\FileLog;
 use CrowdSec\RemediationEngine\CacheStorage\Memcached;
 use CrowdSec\RemediationEngine\CacheStorage\PhpFiles;
 use CrowdSec\RemediationEngine\CacheStorage\Redis;
-use CrowdSec\RemediationEngine\CapiRemediation;
 use CrowdSec\RemediationEngine\Constants as RemConstants;
-use CrowdSecBouncer\AbstractCapiBouncer;
+use CrowdSecBouncer\AbstractBouncerBuilder;
 use CrowdSecBouncer\Bouncer;
 use CrowdSecBouncer\BouncerException;
 use CrowdSecBouncer\Constants;
@@ -35,10 +33,9 @@ use Psr\Log\LoggerInterface;
  *
  * @covers   \CrowdSecBouncer\AbstractBouncer::getRemediationEngine
  * @covers \CrowdSecBouncer\AbstractBouncer::__construct
- * @covers \CrowdSecBouncer\AbstractCapiBouncer::buildClient
- * @covers \CrowdSecBouncer\AbstractCapiBouncer::buildRemediationEngine
- * @covers \CrowdSecBouncer\AbstractCapiBouncer::__construct
- *
+ * @covers \CrowdSecBouncer\AbstractBouncerBuilder::buildClient
+ * @covers \CrowdSecBouncer\AbstractBouncerBuilder::buildRemediationEngine
+ * @covers \CrowdSecBouncer\AbstractBouncerBuilder::__construct
  * @covers   \CrowdSecBouncer\AbstractBouncer::handleCache
  * @covers   \CrowdSecBouncer\AbstractBouncer::handleClient
  * @covers \CrowdSecBouncer\AbstractBouncer::refreshBlocklistCache
@@ -138,7 +135,7 @@ final class AbstractCapiBouncerTest extends TestCase
     {
         $bouncerConfigs = array_merge($this->configs, ['unexpected_config' => 'test']);
         $logger = new FileLog();
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs, $this->storage,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, $this->storage,
             $logger]);
 
         $this->assertEquals(false, $bouncer->getConfig('stream_mode'), 'Stream mode config');
@@ -178,7 +175,7 @@ final class AbstractCapiBouncerTest extends TestCase
             'forced_test_ip' => TestHelpers::BAD_IP,
         ]);
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs, $this->storage], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, $this->storage], '', true,
             true, true, [
                 'sendResponse',
                 'redirectResponse',
@@ -297,7 +294,7 @@ final class AbstractCapiBouncerTest extends TestCase
         // STEP 4 : resolve captcha success
 
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs, $this->storage], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, $this->storage], '', true,
             true, true, [
                 'sendResponse',
                 'redirectResponse',
@@ -344,7 +341,7 @@ final class AbstractCapiBouncerTest extends TestCase
         ]);
 
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs, $this->storage], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, $this->storage], '', true,
             true, true, [
                 'sendResponse',
                 'redirectResponse',
@@ -398,7 +395,7 @@ final class AbstractCapiBouncerTest extends TestCase
 
         // Test 2: not bouncing exclude URI
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$this->configs,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$this->configs,
             $this->storage, $this->logger],
             '', true,
             true, true, [
@@ -423,7 +420,7 @@ final class AbstractCapiBouncerTest extends TestCase
 
         // Test 3: bouncing URI
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$this->configs,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$this->configs,
             $this->storage, $this->logger],
             '', true,
             true, true, [
@@ -443,7 +440,7 @@ final class AbstractCapiBouncerTest extends TestCase
         // Test 4:  not bouncing URI if disabled
         $bouncerConfigs = array_merge($this->configs, ['bouncing_level' => Constants::BOUNCING_LEVEL_DISABLED]);
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs,
             $this->storage, $this->logger],
             '', true,
             true, true, [
@@ -475,7 +472,7 @@ final class AbstractCapiBouncerTest extends TestCase
             ]
         );
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs,
             $this->storage, $this->logger],
             '', true,
             true, true, [
@@ -507,7 +504,7 @@ final class AbstractCapiBouncerTest extends TestCase
             ]
         );
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs,
             $this->storage, $this->logger],
             '', true,
             true, true, [
@@ -534,7 +531,7 @@ final class AbstractCapiBouncerTest extends TestCase
         // Test 9 non-authorized
         $bouncerConfigs = $this->configs;
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs,
             $this->storage, $this->logger],
             '', true,
             true, true, [
@@ -562,7 +559,7 @@ final class AbstractCapiBouncerTest extends TestCase
         // Test 10 authorized
         $bouncerConfigs = $this->configs;
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs,
             $this->storage, $this->logger],
             '', true,
             true, true, [
@@ -591,7 +588,7 @@ final class AbstractCapiBouncerTest extends TestCase
     {
         // handleCache
         $configs = array_merge($this->configs, ['cache_system' => 'redis']);
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$configs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs, $this->storage]);
 
         $result = PHPUnitUtil::callMethod(
             $bouncer,
@@ -602,7 +599,7 @@ final class AbstractCapiBouncerTest extends TestCase
         $this->assertInstanceOf(Redis::class, $result);
 
         $configs = array_merge($this->configs, ['cache_system' => 'memcached']);
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$configs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs, $this->storage]);
 
         $result = PHPUnitUtil::callMethod(
             $bouncer,
@@ -613,7 +610,7 @@ final class AbstractCapiBouncerTest extends TestCase
         $this->assertInstanceOf(Memcached::class, $result);
 
         $configs = array_merge($this->configs, ['cache_system' => 'phpfs']);
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$configs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs, $this->storage]);
 
         $result = PHPUnitUtil::callMethod(
             $bouncer,
@@ -627,7 +624,7 @@ final class AbstractCapiBouncerTest extends TestCase
 
         try {
             $configs = array_merge($this->configs, ['cache_system' => 'phpfs']);
-            $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$configs, $this->storage]);
+            $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs, $this->storage]);
 
             PHPUnitUtil::callMethod(
                 $bouncer,
@@ -647,7 +644,7 @@ final class AbstractCapiBouncerTest extends TestCase
 
         // buildClient
         $configs = $this->configs;
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$configs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs, $this->storage]);
 
         $result = PHPUnitUtil::callMethod(
             $bouncer,
@@ -658,7 +655,7 @@ final class AbstractCapiBouncerTest extends TestCase
         $this->assertEquals('CrowdSec\CapiClient\Client\CapiHandler\FileGetContents', \get_class($result->getRequestHandler()));
 
         $configs = array_merge($this->configs, ['use_curl' => true]);
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$configs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs, $this->storage]);
 
         $result = PHPUnitUtil::callMethod(
             $bouncer,
@@ -677,7 +674,7 @@ final class AbstractCapiBouncerTest extends TestCase
         // Init bouncer
         $bouncerConfigs = $this->configs;
 
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, $this->storage]);
         // Test cache adapter
         $cacheAdapter = $bouncer->getRemediationEngine()->getCacheStorage();
         $cacheAdapter->clear();
@@ -692,11 +689,11 @@ final class AbstractCapiBouncerTest extends TestCase
 
         // Reconfigure the bouncer to set maximum remediation level to "captcha"
         $bouncerConfigs['bouncing_level'] = Constants::BOUNCING_LEVEL_FLEX;
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, $this->storage]);
         $cappedRemediation = $bouncer->getRemediationForIp(TestHelpers::BAD_IP);
         $this->assertEquals('captcha', $cappedRemediation, 'The remediation for the banned IP should now be "captcha"');
         $bouncerConfigs['bouncing_level'] = Constants::BOUNCING_LEVEL_NORMAL;
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, $this->storage]);
         $this->assertEquals(
             'bypass',
             $bouncer->getRemediationForIp(TestHelpers::CLEAN_IP),
@@ -728,7 +725,7 @@ final class AbstractCapiBouncerTest extends TestCase
         );
 
         // Set up a new instance.
-        $bouncer = $this->getMockForAbstractClass(AbstractCapiBouncer::class, [$bouncerConfigs, $this->storage]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, $this->storage]);
 
         $this->assertEquals(
             'ban',