+++ tests/Integration/AbstractLapiBouncerTest.php
@@ -10,8 +10,7 @@ use CrowdSec\Common\Logger\FileLog;
 use CrowdSec\RemediationEngine\CacheStorage\Memcached;
 use CrowdSec\RemediationEngine\CacheStorage\PhpFiles;
 use CrowdSec\RemediationEngine\CacheStorage\Redis;
-use CrowdSec\RemediationEngine\LapiRemediation;
-use CrowdSecBouncer\AbstractLapiBouncer;
+use CrowdSecBouncer\AbstractBouncerBuilder;
 use CrowdSecBouncer\Bouncer;
 use CrowdSecBouncer\BouncerException;
 use CrowdSecBouncer\Constants;
@@ -24,9 +23,9 @@ use Psr\Log\LoggerInterface;
  * @covers \CrowdSecBouncer\AbstractBouncer::clearCache
  * @covers \CrowdSecBouncer\AbstractBouncer::pruneCache
  * @covers \CrowdSecBouncer\AbstractBouncer::testCacheConnection
- * @covers \CrowdSecBouncer\AbstractLapiBouncer::__construct
- * @covers \CrowdSecBouncer\AbstractLapiBouncer::buildClient
- * @covers \CrowdSecBouncer\AbstractLapiBouncer::buildRemediationEngine
+ * @covers \CrowdSecBouncer\AbstractBouncerBuilder::__construct
+ * @covers \CrowdSecBouncer\AbstractBouncerBuilder::buildClient
+ * @covers \CrowdSecBouncer\AbstractBouncerBuilder::buildRemediationEngine
  *
  * @uses   \CrowdSecBouncer\AbstractBouncer::__construct
  * @uses   \CrowdSecBouncer\AbstractBouncer::capRemediationLevel
@@ -65,7 +64,6 @@ use Psr\Log\LoggerInterface;
  * @covers \CrowdSecBouncer\AbstractBouncer::getCache
  *
  * @uses \CrowdSecBouncer\AbstractBouncer::getCaptchaHtml
- *
  *
  * @covers \CrowdSecBouncer\AbstractBouncer::handleCaptchaRemediation
  * @covers \CrowdSecBouncer\AbstractBouncer::handleCaptchaResolutionForm
@@ -153,7 +151,7 @@ final class AbstractLapiBouncerTest extends TestCase
     {
         $bouncerConfigs = array_merge($this->configs, ['unexpected_config' => 'test']);
         $logger = new FileLog();
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs, $logger]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, null, $logger]);
 
         $this->assertEquals(false, $bouncer->getConfig('stream_mode'), 'Stream mode config');
         $this->assertEquals(FileLog::class, \get_class($bouncer->getLogger()), 'Logger Init');
@@ -203,7 +201,7 @@ final class AbstractLapiBouncerTest extends TestCase
         }
 
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs], '', true,
             true, true, [
                 'sendResponse',
                 'redirectResponse',
@@ -319,7 +317,7 @@ final class AbstractLapiBouncerTest extends TestCase
         // STEP 4 : resolve captcha success
 
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs], '', true,
             true, true, [
                 'sendResponse',
                 'redirectResponse',
@@ -376,7 +374,7 @@ final class AbstractLapiBouncerTest extends TestCase
         }
 
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs], '', true,
             true, true, [
                 'sendResponse',
                 'redirectResponse',
@@ -421,7 +419,7 @@ final class AbstractLapiBouncerTest extends TestCase
 
         // Test 2: not bouncing exclude URI
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$this->configs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$this->configs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -445,7 +443,7 @@ final class AbstractLapiBouncerTest extends TestCase
 
         // Test 3: bouncing URI
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$this->configs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$this->configs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -464,7 +462,7 @@ final class AbstractLapiBouncerTest extends TestCase
         // Test 4:  not bouncing URI if disabled
         $bouncerConfigs = array_merge($this->configs, ['bouncing_level' => Constants::BOUNCING_LEVEL_DISABLED]);
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -497,7 +495,7 @@ final class AbstractLapiBouncerTest extends TestCase
             ]
         );
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -544,7 +542,7 @@ final class AbstractLapiBouncerTest extends TestCase
             ]
         );
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -583,7 +581,7 @@ final class AbstractLapiBouncerTest extends TestCase
             ]
         );
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -614,7 +612,7 @@ final class AbstractLapiBouncerTest extends TestCase
             ]
         );
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -640,7 +638,7 @@ final class AbstractLapiBouncerTest extends TestCase
         // Test 9 non-authorized
         $bouncerConfigs = $this->configs;
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -667,7 +665,7 @@ final class AbstractLapiBouncerTest extends TestCase
         // Test 10 authorized
         $bouncerConfigs = $this->configs;
         // Mock sendResponse and redirectResponse to avoid PHP UNIT header already sent or exit error
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs, $this->logger],
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs, null, $this->logger],
             '', true,
             true, true, [
                 'sendResponse',
@@ -695,7 +693,7 @@ final class AbstractLapiBouncerTest extends TestCase
     {
         // handleCache
         $configs = array_merge($this->configs, ['cache_system' => 'redis']);
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$configs], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs], '', true,
             true, true, ['getRemediationEngine']);
 
         $result = PHPUnitUtil::callMethod(
@@ -707,7 +705,7 @@ final class AbstractLapiBouncerTest extends TestCase
         $this->assertInstanceOf(Redis::class, $result);
 
         $configs = array_merge($this->configs, ['cache_system' => 'memcached']);
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$configs], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs], '', true,
             true, true, ['getRemediationEngine']);
 
         $result = PHPUnitUtil::callMethod(
@@ -719,7 +717,7 @@ final class AbstractLapiBouncerTest extends TestCase
         $this->assertInstanceOf(Memcached::class, $result);
 
         $configs = array_merge($this->configs, ['cache_system' => 'phpfs']);
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$configs], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs], '', true,
             true, true, ['getRemediationEngine']);
 
         $result = PHPUnitUtil::callMethod(
@@ -734,7 +732,7 @@ final class AbstractLapiBouncerTest extends TestCase
 
         try {
             $configs = array_merge($this->configs, ['cache_system' => 'phpfs']);
-            $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$configs], '', true,
+            $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs], '', true,
                 true, true, ['getRemediationEngine']);
 
             PHPUnitUtil::callMethod(
@@ -755,25 +753,25 @@ final class AbstractLapiBouncerTest extends TestCase
 
         // buildClient
         $configs = $this->configs;
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$configs], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs], '', true,
             true, true, ['getRemediationEngine']);
 
         $result = PHPUnitUtil::callMethod(
             $bouncer,
             'buildClient',
-            [$configs, new FileLog()]
+            [$configs, null, new FileLog()]
         );
 
         $this->assertEquals(FileGetContents::class, \get_class($result->getRequestHandler()));
 
         $configs = array_merge($this->configs, ['use_curl' => true]);
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$configs], '', true,
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$configs], '', true,
             true, true, ['getRemediationEngine']);
 
         $result = PHPUnitUtil::callMethod(
             $bouncer,
             'buildClient',
-            [$configs, new FileLog()]
+            [$configs, null, new FileLog()]
         );
 
         $this->assertEquals(Curl::class, \get_class($result->getRequestHandler()));
@@ -801,7 +799,7 @@ final class AbstractLapiBouncerTest extends TestCase
             $this->addTlsConfig($bouncerConfigs, $this->useTls);
         }
 
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
 
         // Test cache adapter
         $cacheAdapter = $bouncer->getRemediationEngine()->getCacheStorage();
@@ -840,12 +838,12 @@ final class AbstractLapiBouncerTest extends TestCase
 
         // Reconfigure the bouncer to set maximum remediation level to "captcha"
         $bouncerConfigs['bouncing_level'] = Constants::BOUNCING_LEVEL_FLEX;
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
         $cappedRemediation = $bouncer->getRemediationForIp(TestHelpers::BAD_IP);
         $this->assertEquals('captcha', $cappedRemediation, 'The remediation for the banned IP should now be "captcha"');
         // Reset the max remediation level to its origin state
         $bouncerConfigs['bouncing_level'] = Constants::BOUNCING_LEVEL_NORMAL;
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
 
         $this->logger->info('', ['message' => 'set "Large IPV4 range banned" state']);
         $this->watcherClient->deleteAllDecisions();
@@ -915,7 +913,7 @@ final class AbstractLapiBouncerTest extends TestCase
             $this->addTlsConfig($bouncerConfigs, $this->useTls);
         }
 
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
         // Test cache adapter
         $cacheAdapter = $bouncer->getRemediationEngine()->getCacheStorage();
         $cacheAdapter->clear();
@@ -932,11 +930,11 @@ final class AbstractLapiBouncerTest extends TestCase
 
         // Reconfigure the bouncer to set maximum remediation level to "captcha"
         $bouncerConfigs['bouncing_level'] = Constants::BOUNCING_LEVEL_FLEX;
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
         $cappedRemediation = $bouncer->getRemediationForIp(TestHelpers::BAD_IP);
         $this->assertEquals('captcha', $cappedRemediation, 'The remediation for the banned IP should now be "captcha"');
         $bouncerConfigs['bouncing_level'] = Constants::BOUNCING_LEVEL_NORMAL;
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
         $this->assertEquals(
             'bypass',
             $bouncer->getRemediationForIp(TestHelpers::CLEAN_IP),
@@ -985,7 +983,7 @@ final class AbstractLapiBouncerTest extends TestCase
             $bouncerConfigs['tls_verify_peer'] = true;
         }
 
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
 
         $this->assertEquals(
             'ban',