+++ tests/Integration/GeolocationTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CrowdSecBouncer\Tests\Integration;
 
-use CrowdSecBouncer\AbstractLapiBouncer;
+use CrowdSecBouncer\AbstractBouncerBuilder;
 use CrowdSecBouncer\Constants;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
@@ -30,10 +30,9 @@ use Psr\Log\LoggerInterface;
  * @uses \CrowdSecBouncer\Configuration::cleanConfigs
  * @uses \CrowdSecBouncer\Configuration::getConfigTreeBuilder
  * @uses \CrowdSecBouncer\AbstractBouncer::clearCache
- * @uses \CrowdSecBouncer\AbstractLapiBouncer::__construct
- * @uses \CrowdSecBouncer\AbstractLapiBouncer::buildClient
- * @uses \CrowdSecBouncer\AbstractLapiBouncer::buildRemediationEngine
- *
+ * @uses \CrowdSecBouncer\AbstractBouncerBuilder::__construct
+ * @uses \CrowdSecBouncer\AbstractBouncerBuilder::buildClient
+ * @uses \CrowdSecBouncer\AbstractBouncerBuilder::buildRemediationEngine
  */
 final class GeolocationTest extends TestCase
 {
@@ -128,7 +127,7 @@ final class GeolocationTest extends TestCase
             'stream_mode' => false,
         ]);
 
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
 
         $bouncer->clearCache();
 
@@ -148,7 +147,7 @@ final class GeolocationTest extends TestCase
         $geolocationConfig['enabled'] = false;
         $bouncerConfigs['geolocation'] = $geolocationConfig;
 
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
         $bouncer->clearCache();
 
         $this->assertEquals(
@@ -161,7 +160,7 @@ final class GeolocationTest extends TestCase
         $this->watcherClient->setSecondState();
         $geolocationConfig['enabled'] = true;
         $bouncerConfigs['geolocation'] = $geolocationConfig;
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
         $bouncer->clearCache();
 
         $this->assertEquals(
@@ -200,7 +199,7 @@ final class GeolocationTest extends TestCase
             'fs_cache_path' => TestHelpers::PHP_FILES_CACHE_ADAPTER_DIR,
         ]);
 
-        $bouncer = $this->getMockForAbstractClass(AbstractLapiBouncer::class, [$bouncerConfigs]);
+        $bouncer = $this->getMockForAbstractClass(AbstractBouncerBuilder::class, [$bouncerConfigs]);
         $cacheAdapter = $bouncer->getRemediationEngine()->getCacheStorage();
         $cacheAdapter->clear();